Move implementations to AbstractConnector
